Interdepartmental Proccessing
Her is my code to compute the interdepartmental pre and co requisites
along with the number of course in every department.

I still need to implement the correct non numerical names of
departments, because those were not present in the raw data.

I know that Scotty labs provides this data, but it has not been scraped
yet (as far as I know.)